S01-T10

See merge request Jonny__/jborn_web_10_dzhenenko!18

+++ src/main/java/ru/dzhenenko/entity/User.java
@@ -2,6 +2,8 @@ package ru.dzhenenko.entity;
 
 import lombok.Getter;
 import lombok.Setter;
+import org.springframework.stereotype.Service;
+import ru.dzhenenko.api.json.AuthRequest;
 
 import javax.persistence.*;
 import java.util.List;
@@ -18,7 +20,8 @@ public class User {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+
+    private long id;
 
     @Column(name = "first_name")
     private String firstName;
